adaptive_rag/pyproject.toml
[project]
name = "adaptive-rag"
version = "0.1.0"
description = "Adaptive RAG for Korean Competition Law Cases"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "fire>=0.7.1",
    "pypdf>=6.10.2",
    "langchain>=1.2.15",
    "langchain-core>=1.3.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.2",
    "langchain-chroma>=1.1.0",
    "langchain-openai>=1.1.14",
    "llama-parse>=0.6.94",
    "unstructured[pdf]",
    "langchain-tavily>=0.1.0",
]
